requirements.txt
attrs==19.3.0
cattrs==1.0.0
certifi==2020.4.5.2
chardet==3.0.4
future==0.18.2
idna==2.9
oauthlib==3.1.0
python-dotenv==0.13.0
python-facebook-api==0.6.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
typing==3.7.4.1
urllib3==1.25.9
